Comment out BMI calculation code
The BMI calculation part of the code has been commented out. More specifically, the instantiation of the IBMICalculatorStrategy, the BMICategoryInterpreter classes, and the BMI calculation logic have been disabled. This change was due to an ongoing redesign of the BMI calculation process which presently causes an error during compilation. The changes are in line with a TODO task to reactivate these segments of code once the new design is implemented.